Task 9.2. updated
