+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/Application.java
@@ -1,5 +1,12 @@
 package com.kodilla.good.patterns.challenges;
 
+import com.kodilla.good.patterns.challenges.information.MailService;
+import com.kodilla.good.patterns.challenges.order.OrderProcessor;
+import com.kodilla.good.patterns.challenges.order.OrderRequest;
+import com.kodilla.good.patterns.challenges.order.OrderRequestRetriever;
+import com.kodilla.good.patterns.challenges.repository.CarOrderRepository;
+import com.kodilla.good.patterns.challenges.service.CarOrderService;
+
 public class Application {
     public static void main(String[] args) {
         OrderRequestRetriever orderRequestRetriever = new OrderRequestRetriever();
