--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/CarOrderRepository.java
@@ -1,8 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-public class CarOrderRepository implements OrderRepository {
-    @Override
-    public void createOrder(User user, Item item) {
-
-    }
-}
