--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/CarOrderService.java
@@ -1,8 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-public class CarOrderService implements OrderService{
-    @Override
-    public boolean order(User user, Item item) {
-        return true;
-    }
-}
