--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/OrderService.java
@@ -1,5 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-public interface OrderService {
-    boolean order(User user, Item item);
-}
