+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/information/InformationService.java
@@ -1,0 +1,7 @@
+package com.kodilla.good.patterns.challenges.information;
+
+import com.kodilla.good.patterns.challenges.models.User;
+
+public interface InformationService {
+    void inform(User user);
+}
