+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/models/Item.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.models;
 
 public class Item {
     private String itemName;
