+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/models/User.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.models;
 
 public class User {
     private String name;
