+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/moviestore/MovieStore.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.moviestore;
 
 import java.util.ArrayList;
 import java.util.HashMap;
