+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/order/OrderDto.java
@@ -1,4 +1,6 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.order;
+
+import com.kodilla.good.patterns.challenges.models.User;
 
 public class OrderDto {
     public User user;
