+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/service/OrderService.java
@@ -1,0 +1,8 @@
+package com.kodilla.good.patterns.challenges.service;
+
+import com.kodilla.good.patterns.challenges.models.Item;
+import com.kodilla.good.patterns.challenges.models.User;
+
+public interface OrderService {
+    boolean order(User user, Item item);
+}
