+++ settings.gradle
@@ -1,4 +1,5 @@
 rootProject.name = 'kodilla-course'
 include 'kodilla-testing'
 include 'kodilla-stream'
-
+include 'kodilla-good-patterns'
+include 'kodilla-exception'
